pyproject.toml
[project]
name = "ip-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "grequests>=0.7.0",
    "python-dotenv>=1.0.1",
]
